% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnaSeqTutorials-package.R
\docType{package}
\name{RnaSeqTutorial04 package}
\alias{RnaSeqTutorial04}
\alias{RnaSeqTutorial04-package}
\alias{RnaSeqTutorial04 package}
\title{RnaSeqTutorials package}
\description{
A simple description of the RnaSeqTutorial04 package
}
\section{Tutorials}{

This is the fourth in a series of tutorials
\itemize{
\item\code{04_gene_set_enrichment_analysis} a gene set enrichment analysis
}
}

\examples{
	\dontrun{
	learnr::run_tutorial("04_gene_set_enrichment_analysis", package = "RnaSeqTutorial04")
	}
}
\seealso{
The vignette
}
\author{
Nicolas Delhomme [aut,cre]
}
\keyword{internal}
\keyword{package}
